GraphAnalysis
This code allows to identify users within a credit organisation who need to be investigated based on the performance indicators of loans that they work with. The code requires loan data with user ids, NPF indicators, loan balance, credited amount, non-serviced indicator, and less4months indicators linked to the client of the loan.
The code can be re-used for any dataset where there is a list of user id's associated with some action (edges of graph), and a list of 'goodness' of action needs to be weighted to show the 'quality' of connection between users.